requirements.txt
# Python Version: 3.10.11
streamlit==1.40.0
pandas==1.5.3
torch==2.5.0
joblib==1.3.1
numpy==1.24.3
mysql-connector-python==9.0.0
scikit-learn==1.2.2
pymongo==4.9.1
python-dotenv==1.0.1